# openapi.yaml
openapi: 3.1.0
info:
  title: NextGesAPI
  description: API for NextGes
  summary: This documentation provides an overview of the NextGes API, including authentication and entity operations.
  version: 1.0.0
  contact:
    name: NextGes Support
    email: tristan.deloeil@gmail.com
    url: https://flendergit.github.io
  x-logo:
    url: https://www.svgrepo.com/show/54409/school.svg
    altText: NextGes Logo
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms
...
paths:
  /login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate a user and retrieve an access token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  format: password
                  description: User's password
            examples:
              example1:
                summary: Example login request
                value:
                  email: "exemple@test.com"
                  password: "password123"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: Access token for authenticated requests
                  jwt:
                    type: string
                    description: JWT token for authentication
              examples:
                example1:
                  summary: Successful login response
                  value:
                    access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    jwt: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /me:
    get:
      tags:
        - Authentication
      summary: Get current user information
      description: Retrieve information about the authenticated user
      operationId: getCurrentUser
      responses:
        "200":
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claims"
              examples:
                example1:
                  summary: Successful user information response
                  value:
                    sub: "1234567890"
                    name: "John Doe"
                    email: "john.doe@test.com"
                    scopes:
                      - "read:users"
                      - "write:users"
                    exp: 1700000000
                    iat: 1600000000
                    iss: "NextGes API"
                    aud: "NextGes"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            example1:
              summary: Unauthorized error
              value:
                trace-id: "12345"
                message: "Unauthorized access"
                timestamp: "2023-10-01T12:00:00Z"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            example1:
              summary: Forbidden error
              value:
                trace-id: "12345"
                message: "Access denied"
                timestamp: "2023-10-01T12:00:00Z"
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            example1:
              summary: Internal Server Error
              value:
                trace-id: "12345"
                message: "An unexpected error occurred"
                timestamp: "2023-10-01T12:00:00Z"
  schemas:
...
    Claims:
      type: object
      properties:
        sub:
          type: string
          description: Subject of the token (user ID)
        name:
          type: string
          description: Name of the user
        email:
          type: string
          format: email
          description: Email address of the user
        scopes:
          type: array
          items:
            type: string
          description: List of scopes associated with the user
        exp:
          type: integer
          format: int64
          description: Expiration time of the token (in seconds since epoch)
        iat:
          type: integer
          format: int64
          description: Issued at time of the token (in seconds since epoch)
        iss:
          type: string
          description: Issuer of the token (e.g., NextGes API)
        aud:
          type: string
          description: Audience of the token (e.g., intended recipient)
...
    ErrorResponse:
      type: object
      properties:
        trace-id:
          type: string
          description: Unique identifier for the error trace
        message:
          type: string
          description: Error message
        timestamp:
          type: string
          format: date-time
          description: Date and time when the error occurred
      required:
        - trace-id
        - message
        - timestamp